chore: reorganize mod file
